Add 'conns_test.go'.
